♻️ refactor(docker_file): refactored builder image by adding a directory folder and generating prisma before building. Also allowed using prisma with alkine version of node image by adding the binaryTargets inside the schema.prisma file. Also removed any unused vars and added missed type decleration to be able to build the production version of our app so we finally could build it using docker

+++ src/modules/comments/controllers/deleteComment.ts
@@ -3,6 +3,7 @@ import { handlePrismaError } from "../../../utils/helperMethods/handlePrismaErro
 import { prisma } from "../../../services/prismaClient.js";
 import { getJsonResponse } from "../../../utils/helperMethods/getJsonResponse.js";
 import { getSuccessMsg } from "../../../utils/helperMethods/generateSuccessMsg.js";
+import { getAuth } from "@clerk/express";
 
 export async function deleteComment(
   req: Request,
@@ -10,10 +11,25 @@ export async function deleteComment(
   next: NextFunction
 ) {
   try {
+    const { userId } = getAuth(req);
+    if (!userId) throw new Error("UserId from clerk is not found!!");
+    const user = await prisma.user.findUnique({
+      where: {
+        clerkId: userId,
+      },
+    });
+    if (!user) throw new Error("User not found");
+
     const commentId = req.params.id;
+
+    const postSlug = req.query.postSlug?.toString();
+    if (!postSlug) return next(new Error("Post slug is required."));
+
     const deletedComment = await prisma.comment.delete({
       where: {
         id: commentId,
+        userEmail: user.email,
+        postSlug,
       },
       include: {
         user: {
